making sure swarm key is persistent when universe is saved + consolidating some of the swam key generation logic
